+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 ## [Unreleased]
 
+### Changed
+
+- The config is now required:
+	- `eslint` version `9.39.2` or higher.
+	- `node.js` version `20.19.0` or higher.
+
 ## [5.0.0] — 2024–11–16
 
 ### Changed